services population duration issue fixed

+++ application/controllers/AccountController.php
@@ -319,8 +319,11 @@ class AccountController extends Zend_Controller_Action implements Aimya_Controll
 
                 $dbProfile = new Application_Model_DbTable_Profile();
                 $profileModel = new Application_Model_Profile();
-                unlink(substr($profileModel->getAvatarPath($identity->id, 'base'), 1)); //substr first slah
-                unlink(substr($profileModel->getAvatarPath($identity->id, 'medium'), 1));
+                if(substr($profileModel->getAvatarPath($identity->id, 'base'), 0, 12)!='/img/design/'){
+                    unlink(substr($profileModel->getAvatarPath($identity->id, 'base'), 1)); //substr first slah
+                    unlink(substr($profileModel->getAvatarPath($identity->id, 'medium'), 1));
+                }
+
 
                 $dbProfile->deleteAvatar($identity->id);
